Feat: meta 테이블을 enum으로 활용하는 기능 추가
