+++ src/main/java/kr/sesacjava/swimtutor/routine/dto/RequestTrainingForRoutineDTO.java
@@ -1,0 +1,19 @@
+package kr.sesacjava.swimtutor.routine.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@ToString
+@AllArgsConstructor
+@NoArgsConstructor
+public class RequestTrainingForRoutineDTO {
+    // TrainingForRoutine
+    private String session;
+
+    // Training
+    private String trainingId;
+    private String strokeName;
+    private int distance;
+    private int sets;
+}
